refactor: Sweep dead code and stale transport comments after socket unification
